Moved and renamed DraftResponsible to Responsible in common
