+++ concept-api/src/main/scala/no/ndla/conceptapi/model/api/Concept.scala
@@ -30,6 +30,6 @@ case class Concept(
     @(ApiModelProperty @field)(description = "Article ids to which the concept is connected to") articleIds: Seq[Long],
     @(ApiModelProperty @field)(description = "Status information of the concept") status: Status,
     @(ApiModelProperty @field)(description = "A visual element for the concept") visualElement: Option[VisualElement],
-    @(ApiModelProperty @field)(description = "Object with data containing representing the editor responsible for this concept") responsible: Option[ConceptResponsible]
+    @(ApiModelProperty @field)(description = "Object with data representing the editor responsible for this concept") responsible: Option[ConceptResponsible]
     // format: on
 )
