+++ concept-api/src/main/scala/no/ndla/conceptapi/model/domain/Concept.scala
@@ -7,7 +7,7 @@
 
 package no.ndla.conceptapi.model.domain
 
-import no.ndla.common.model.domain.{Tag, Title}
+import no.ndla.common.model.domain.{Responsible, Tag, Title}
 import no.ndla.common.errors.ValidationException
 import no.ndla.common.model.domain.draft.Copyright
 import no.ndla.conceptapi.Props
@@ -37,7 +37,7 @@ case class Concept(
     articleIds: Seq[Long],
     status: Status,
     visualElement: Seq[VisualElement],
-    responsible: Option[ConceptResponsible]
+    responsible: Option[Responsible]
 ) {
 
   lazy val supportedLanguages: Set[String] =
