+++ draft-api/src/main/scala/no/ndla/draftapi/model/api/Article.scala
@@ -43,6 +43,6 @@ case class Article(
     @(ApiModelProperty @field)(description = "Value that dictates who gets to see the article. Possible values are: everyone/teacher") availability: String,
     @(ApiModelProperty @field)(description = "A list of content related to the article") relatedContent: Seq[RelatedContent],
     @(ApiModelProperty @field)(description = "A list of revisions planned for the article") revisions: Seq[RevisionMeta],
-    @(ApiModelProperty @field)(description = "Object with data containing representing the editor responsible for this article") responsible: Option[DraftResponsible],
+    @(ApiModelProperty @field)(description = "Object with data representing the editor responsible for this article") responsible: Option[DraftResponsible],
     @(ApiModelProperty @field)(description = "The path to the frontpage article") slug: Option[String],
 )
